+++ _templates/component/new/component.ejs
@@ -5,7 +5,7 @@ to: src/components/<%= h.inflection.undasherize(name) %>/<%= h.inflection.undash
 import * as React from "react";
 import classNames from "classnames";
 
-import * as styles from "./<%= comp %>.scss";
+import * as styles from "./<%= comp %>.module.scss";
 
 interface Props {
     style?: React.CSSProperties;
