+++ _templates/component/new/style.ejs
@@ -1,5 +1,5 @@
 ---
-to: src/components/<%= h.inflection.undasherize(name) %>/<%= h.inflection.undasherize(name) %>.scss
+to: src/components/<%= h.inflection.undasherize(name) %>/<%= h.inflection.undasherize(name) %>.module.scss
 ---
 .component {
