+++ _templates/component/new/styletyping.ejs
@@ -1,4 +1,6 @@
 ---
-to: src/components/<%= h.inflection.undasherize(name) %>/<%= h.inflection.undasherize(name) %>.scss.d.ts
+to: src/components/<%= h.inflection.undasherize(name) %>/<%= h.inflection.undasherize(name) %>.module.scss.d.ts
 ---
+/* tslint:disable */
+// This file is automatically generated from your CSS. Any edits will be overwritten.
 export const component: string;
