+++ src/components/Card/Card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import * as styles from "./Card.scss";
+import * as styles from "./Card.module.scss";
 
 interface Props {
     style?: React.CSSProperties;
