+++ src/components/FullScreenMedia.module.scss.d.ts
