+++ src/components/HalfBlock/HalfBlock.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import * as styles from "./HalfBlock.scss";
+import * as styles from "./HalfBlock.module.scss";
 
 interface Props {
     style?: React.CSSProperties;
