+++ src/components/PageNavBar/PageNavBar.tsx
@@ -5,7 +5,7 @@ import { PageListNode } from "../../types/data";
 import { MainNavigation } from "../MainNavigation";
 
 import logo from "../../../static/assets/furniture/aerian-logo.svg";
-import * as styles from "./PageNavBar.scss";
+import * as styles from "./PageNavBar.module.scss";
 
 interface Props {
     style?: React.CSSProperties;
