+++ src/components/SocialLinks/SocialLinks.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import * as styles from "./SocialLinks.scss";
+import * as styles from "./SocialLinks.module.scss";
 
 interface Props {
     style?: React.CSSProperties;
