+++ src/components/Tag/Tag.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import * as styles from "./Tag.scss";
+import * as styles from "./Tag.module.scss";
 
 interface Props {
     style?: React.CSSProperties;
